+++ Cargo.toml
@@ -14,3 +14,10 @@ maintenance = { status = "actively-developed" }
 
 [dependencies]
 chrono = { version = "0.4", features = ["serde"] }
+
+[dev-dependencies]
+criterion = "0.2"
+
+[[bench]]
+name = "b_instruments"
+harness = false
