+++ src/instruments/order.rs
@@ -4,33 +4,30 @@ use std::cell::RefCell;
 
 #[derive(Debug, Eq, PartialEq, Ord, PartialOrd, Clone)]
 pub struct Order {
-    // REVIEW: id
+    // REVIEW: id field?
     pub symbol: &'static str,
     pub price: Currency,
     pub qty: u32,
     pub side: Side,
     pub datetime: DateTime<Utc>,
-    state: RefCell<State>,
+    state: State,
 }
 
 impl Order {
     pub fn new(
         symbol: &'static str,
-        price: f64,
+        price: f32,
         qty: u32,
         side: Side,
-        dt_time: Option<DateTime<Utc>>,
+        datetime: DateTime<Utc>,
     ) -> Order {
         Order {
             symbol,
             price: Currency::new(price),
             qty,
             side,
-            datetime: match dt_time {
-                Some(val) => val,
-                None => Utc::now(),
-            },
-            state: RefCell::new(State::New),
+            datetime,
+            state: State::New,
         }
     }
 }
